Separated mock api and concrete api
