+++ Film/Network/API/MovieInfoAPI.swift
@@ -1,0 +1,13 @@
+//
+//  MockMovieAPI.swift
+//  Film
+//
+//  Created by Michel Balamou on 2019-09-15.
+//  Copyright © 2019 ElasticPanda. All rights reserved.
+//
+
+import Foundation
+
+protocol MovieInfoAPI {
+    func getMovieInfo(movieId: Int, result: @escaping Handler<Movie>)
+}
