+++ Film/Network/API/WatchedAPI.swift
@@ -39,25 +39,3 @@ class ConcreteWatchedAPI: WatchedAPI {
         })
     }
 }
-
-class MockWatchedAPI: WatchedAPI {
-    
-    var count = 0
-    let timeDelay = 1.5
-    
-    func getWatched(result: @escaping WatchedHandler) {
-        DispatchQueue.main.asyncAfter(deadline: .now() + timeDelay) { [weak self] in
-            guard let self = self else { return }
-            
-            if self.count == 2 || self.count == 5 {
-                result(.failure(ConnectionError.invalidURL))
-            } else {
-                let data = (self.count == 1 || self.count == 4) ? [] : Watched.getRandomMock()
-                result(.success(data))
-            }
-            
-            self.count += 1
-        }
-    }
-}
-
